% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{package}
\name{tsvio-package}
\alias{tsvio-package}
\title{tsvio: Simple Utilities for Tab Separated Value (TSV) Files}
\description{
Simple Utilities for Tab Separated Value (TSV) Files
}
\details{
Utilities for indexing and rapidly loading (subsets of) data from (large) tab separated value (TSV) files.
The TSV files are required to have a unique row label in the first column of each line and a unique
column label in the first line of the file.
Files may be formatted in either spreadsheet/Unix format (same number of fields on each line) or R format
(one less column on the first line only).
The data matrix in the files are expected to have the same data types in all entries.  (The row and column
labels are always expected to be strings.)
}
\seealso{
tsvGenIndex, tsvGetLines, tsvGetData
}
